# src/ldfd/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Linux Distribution Factory REST API - Build, manage, and distribute custom Linux distributions.",
        "title": "LDF API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8443",
    "basePath": "/",
    "paths": {
...
        "/v1/language-packs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all available custom language packs with metadata",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LanguagePacks"
                ],
                "summary": "List language packs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/langpacks.LanguagePackListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Uploads a new language pack from a .tar.xz, .tar.gz, or .xz archive containing meta.json and translation files",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LanguagePacks"
                ],
                "summary": "Upload language pack",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Language pack archive (.tar.xz, .tar.gz, or .xz)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Language pack updated",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "201": {
                        "description": "Language pack created",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/language-packs/{locale}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a specific language pack including its full translation dictionary",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LanguagePacks"
                ],
                "summary": "Get language pack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Locale identifier (e.g. fr, de, ja)",
                        "name": "locale",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/langpacks.LanguagePackResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a language pack by its locale identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LanguagePacks"
                ],
                "summary": "Delete language pack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Locale identifier (e.g. fr, de, ja)",
                        "name": "locale",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "error": {
                    "type": "string",
                    "example": "Not found"
                },
                "message": {
                    "type": "string",
                    "example": "The requested resource was not found"
                }
            }
        },
...
        "db.LanguagePackMeta": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "locale": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "langpacks.LanguagePackListResponse": {
            "type": "object",
            "properties": {
                "language_packs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.LanguagePackMeta"
                    }
                }
            }
        },
        "langpacks.LanguagePackResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "dictionary": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "locale": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token authentication. Prefix the token with \"Bearer \".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}